refactor: props 제거 및 type->interface.ts
