+++ src/pages/MapPage/interface.ts
@@ -1,43 +1,5 @@
 import { Slope } from '../../apis/slopeMap';
 
-export interface MapComponentProps {
-  selectedMarkerId: number | null;
-  escarpmentData: Slope[];
-  allTextShow: boolean;
-  userLocation: naver.maps.LatLng | null;
-  setUserLocation: (location: naver.maps.LatLng | null) => void;
-  mapInstance: naver.maps.Map | null;
-  setMapInstance: (map: naver.maps.Map | null) => void;
-  onMarkerClick: (item: Slope, index: number) => void;
-}
-export interface BottomSheetProps {
-  slopeData: Slope[];
-  selectItem: Slope | null;
-  onItemClick: (item: Slope, index: number) => void;
-  height: number;
-  setHeight: (height: number) => void;
-  onCloseInfo: () => void;
-  searchMod: boolean;
-}
-export interface InfotableProps {
-  selectItem: Slope | null;
-  onCloseInfo: () => void;
-}
-export interface ListProps {
-  item: Slope | null;
-  onClick?: () => void;
-}
-
-export interface SearchComponentProps {
-  onSearch: (value: string) => void;
-}
-
-interface UserInfo {
-  _id: string;
-  name: string;
-  organization: string;
-  isAdmin: boolean;
-}
 export interface CommentData {
   _id: string;
   slopeId: string;
@@ -48,7 +10,76 @@ export interface CommentData {
   updatedAt: string;
   __v: number;
 }
+interface UserInfo {
+  _id: string;
+  name: string;
+  organization: string;
+  isAdmin: boolean;
+}
+
+//props관련
+export interface InfotableProps {
+  selectItem: Slope | null;
+}
+export interface ListProps {
+  item: Slope | null;
+  onClick?: () => void;
+}
+
+export interface SearchComponentProps {
+  onSearch: (value: string) => void;
+}
+export interface NoInfoProps {
+  text: string;
+}
 export interface CommentContainerProps {
   comment: CommentData;
   fetchComment: () => Promise<void>;
+}
+export interface SearchResultProps {
+  resultCount: number;
+  gradeCount: {
+    A: number;
+    B: number;
+    C: number;
+    D: number;
+    F: number;
+  };
+}
+export interface DeleteModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+  onSubmit: () => void;
+}
+export interface CommentListProps {
+  slopeId: string;
+}
+
+//모달 관련
+
+export interface DeleteConfirmModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+}
+
+export interface UserDelete {
+  _id: string;
+  isAdmin: string;
+  name: string;
+  organization: string;
+  phone: string;
+}
+
+export interface LeftModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+}
+
+export interface PrivacyPolicyModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+}
+export interface TermsofUseModalProps {
+  isOpen: boolean;
+  onClose: () => void;
 }
